make session security context available for both websocket protocols
